+++ appinfo/routes.php
@@ -24,7 +24,7 @@ return [
         ['name' => 'page#publicSessionWatch', 'url' => '/publicSessionWatch/{publicviewtoken}', 'verb' => 'GET'],
         ['name' => 'page#createSession', 'url' => '/createSession', 'verb' => 'POST'],
         ['name' => 'page#deleteSession', 'url' => '/deleteSession', 'verb' => 'POST'],
-        ['name' => 'page#deletePoint', 'url' => '/deletePoint', 'verb' => 'POST'],
+        ['name' => 'page#deletePoints', 'url' => '/deletePoints', 'verb' => 'POST'],
         ['name' => 'page#updatePoint', 'url' => '/updatePoint', 'verb' => 'POST'],
         ['name' => 'page#addPoint', 'url' => '/addPoint', 'verb' => 'POST'],
         ['name' => 'page#getSessions', 'url' => '/getSessions', 'verb' => 'POST'],